requirements.txt
mlflow
scikit-learn
pandas
numpy
azure-ai-ml
azure-identity
joblib
